% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov95_table.R
\name{cov95_table}
\alias{cov95_table}
\title{Creates a 95\% prediction interval coverage table}
\usage{
cov95_table(wis_data)
}
\arguments{
\item{wis_data}{Data frame of results generated from 'WIS_func' function}
}
\value{
Table of 95\% prediction interval coverage
}
\description{
Creates a 95\% prediction interval coverage table
}
