% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_for_scores_function.R
\name{dat_for_scores_function}
\alias{dat_for_scores_function}
\title{Data frame for scoring}
\usage{
dat_for_scores_function(all_dat, obs_data)
}
\arguments{
\item{all_dat}{Projection (outcome) data; all_dat must include columns: target_end_date, forecast_date, quantile, value, type, location_name, location, and age_group}

\item{obs_data}{"Ground truth" (observed) data; obs_dat must include columns: value_inc, target_end_date, location_name, location, and age_group}
}
\value{
Data frame with observed and projection data - ready to be scored
}
\description{
Data frame for scoring
}
