% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mean_model_output}
\alias{example_mean_model_output}
\title{Example model output data representing mean point forecasts}
\format{
\subsection{\code{example_mean_model_output}}{

A data frame with 16 rows and 8 columns:
\describe{
\item{origin_date}{Date: Reference date for the forecast.}
\item{horizon}{integer: Forecast horizon, in days.}
\item{location}{character: FIPS codes identifying the location being forecasted.}
\item{target}{character: The forecast target. All values are "inc covid hosp".}
\item{output_type}{character: Output type used to represent the forecast. All values are "mean".}
\item{output_type_id}{NA: This column is not used when the output type is "mean". All values are "NA".}
\item{value}{numeric: The mean of the forecast distribution.}
\item{model_id}{character: Identifier for the model that produced the forecast.}
}
}
}
\source{
\url{https://github.com/reichlab/covid19-forecast-hub}
}
\usage{
example_mean_model_output
}
\description{
These are based on a subset of forecasts of daily incident hospital
admissions for COVID-19 that were submitted to the US COVID-19 Forecast hub,
which have been lightly modified for expository purposes. To obtain point
forecasts, we used the \code{distfromq} package for R to estimate the quantile
function, drew 100,000 samples from the distribution using inverse transform
sampling, and computed the mean of those samples.
}
\keyword{datasets}
