% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_quantile_model_output}
\alias{example_quantile_model_output}
\title{Example model output data representing quantile forecasts}
\format{
\subsection{\code{example_quantile_model_output}}{

A data frame with 112 rows and 8 columns:
\describe{
\item{origin_date}{Date: Reference date for the forecast.}
\item{horizon}{integer: Forecast horizon, in days.}
\item{location}{character: FIPS codes identifying the location being forecasted.}
\item{target}{character: The forecast target. All values are "inc covid hosp".}
\item{output_type}{character: Output type used to represent the forecast. All values are "quantile".}
\item{output_type_id}{numeric: The probability level for the quantile forecast.
One of 0.05, 0.1, 0.25, 0.5, 0.75, 0.9 or 0.95.}
\item{value}{integer: The quantile of the forecast distribution.}
\item{model_id}{character: Identifier for the model that produced the forecast.}
}
}
}
\source{
\url{https://github.com/reichlab/covid19-forecast-hub}
}
\usage{
example_quantile_model_output
}
\description{
These are a subset of forecasts of daily incident hospital admissions for
COVID-19 that were submitted to the US COVID-19 Forecast hub, which have
been lightly modified for expository purposes.
}
\keyword{datasets}
