% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_target_data}
\alias{example_target_data}
\title{Example target data recording outcomes that are predicted in the
forecasts stored in \code{example_quantile_model_output}.}
\format{
\subsection{\code{example_target_data}}{

A data frame with 8 rows and 5 columns:
\describe{
\item{origin_date}{Date: Reference date for the forecast.}
\item{horizon}{integer: Forecast horizon, in days.}
\item{location}{character: FIPS codes identifying the location being forecasted.}
\item{target}{character: The forecast target. All values are "inc covid hosp".}
\item{value}{integer: The number of reported hospitalizations.}
}
}
}
\source{
\url{https://healthdata.gov/Hospital/COVID-19-Reported-Patient-Impact-and-Hospital-Capa/g62h-syeh/about_data}
}
\usage{
example_target_data
}
\description{
These are daily incident hospital admissions for COVID-19 that were reported
on HealthData.gov. They are organized so as to match the forecast task id
columns in \code{example_quantile_model_output}, so that the observed data can be
joined into the forecast data frame correctly.
}
\keyword{datasets}
