% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_obs_data.R
\name{format_obs_data}
\alias{format_obs_data}
\title{Transform observed data into appropriate format \emph{NOT FINISHED}}
\usage{
format_obs_data(obs_data, proj_data, target1)
}
\arguments{
\item{target1}{}
}
\description{
Transform observed data into appropriate format \emph{NOT FINISHED}
}
