% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_tibble.R
\name{get_data_tibble}
\alias{get_data_tibble}
\title{Get data tibble with all output (projection) data to be evaluated}
\usage{
get_data_tibble(output_path, loc_data, weeks.to.eval)
}
\arguments{
\item{output_path}{File path to folder with output data}

\item{loc_data}{Data file with location information, including variables location and location_name}

\item{weeks.to.eval}{Sequence of dates to be evaluated}
}
\description{
Get data tibble with all output (projection) data to be evaluated
}
