% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_sat.R
\name{get_next_sat}
\alias{get_next_sat}
\title{Get next Saturday}
\usage{
get_next_sat(date)
}
\arguments{
\item{date}{Date forecast begins (forecast_date)}
}
\value{
Date of next Saturday
}
\description{
Get next Saturday
}
