% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_saturday.R
\name{get_next_saturday}
\alias{get_next_saturday}
\title{Get next Saturday}
\usage{
get_next_saturday(date)
}
\arguments{
\item{date}{Date forecast begins (forecast_date)}
}
\value{
Date of next Saturday
}
\description{
Get next Saturday
}
