% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_data.R
\name{get_results_data}
\alias{get_results_data}
\title{Function that outputs a data frame with all of the WIS scores}
\usage{
get_results_data(obs_data, weeks_to_eval, output_path, target1, loc_data)
}
\arguments{
\item{obs_data}{Ground truth data file}

\item{weeks_to_eval}{Sequence of dates to be evaluated}

\item{output_path}{File path to directory where output data are stored}

\item{target1}{Target of interest; for example, 'inc hosp'}

\item{loc_data}{Data file containing location information, including FIPS codes and location names}
}
\description{
Function that outputs a data frame with all of the WIS scores
}
