% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_rankings_func.R
\name{inc_rankings_func}
\alias{inc_rankings_func}
\title{Incidence rankings}
\usage{
inc_rankings_func(WIS_data)
}
\arguments{
\item{WIS_data}{Data frame with WIS scoring}
}
\description{
Incidence rankings
}
