% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_point_model_output.R
\name{transform_point_model_output}
\alias{transform_point_model_output}
\title{Transform mean/median model output into a forecast object}
\usage{
transform_point_model_output(proj_data, obs_data, type)
}
\arguments{
\item{proj_data}{Model output projection data frame;
requires columns: model_id, target, value, location, target_end_date, horizon, output_type}

\item{obs_data}{Observed 'ground truth' data to be compared against forecasts;
requires columns: target, observation, location, target_end_date, output_type}

\item{type}{Forecast output type: "mean" or "median"}
}
\value{
forecast_point
}
\description{
Transform mean/median model output into a forecast object
}
