% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quantil_model_output.R
\name{transform_quantil_model_output}
\alias{transform_quantil_model_output}
\title{Transform quantile model output into a forecast object}
\usage{
transform_quantil_model_output(proj_data, obs_data)
}
\arguments{
\item{proj_data}{Model output projection data frame;
requires columns: model_id, target, value, location, target_end_date, horizon, output_type, output_type_id}

\item{obs_data}{Observed 'ground truth' data to be compared against forecasts;
requires columns: target, observation, location, target_end_date, output_type}
}
\value{
forecast_quantile
}
\description{
Transform quantile model output into a forecast object
}
