% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quantile_model_output.R
\name{transform_quantile_model_output}
\alias{transform_quantile_model_output}
\title{Transform quantile model output into a forecast object}
\usage{
transform_quantile_model_output(proj_data, obs_data)
}
\arguments{
\item{proj_data}{Model output projection data frame}

\item{obs_data}{Observed 'ground truth' data to be compared against forecasts}
}
\value{
forecast_quantile
}
\description{
Transform quantile model output into a forecast object
}
