% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIS_func.R
\name{WIS_func}
\alias{WIS_func}
\title{WIS}
\usage{
WIS_func(dat_for_scores, weeks.to.eval)
}
\arguments{
\item{dat_for_scores}{Data frame containing data to be scored}

\item{weeks.to.eval}{}
}
\description{
WIS
}
