% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_model_out.R
\name{score_model_out}
\alias{score_model_out}
\title{Score model output predictions against observed data}
\usage{
score_model_out(
  model_out_tbl,
  target_observations,
  metrics = NULL,
  by = NULL,
  output_type_id_order = NULL
)
}
\arguments{
\item{model_out_tbl}{Model output tibble with predictions}

\item{target_observations}{Observed 'ground truth' data to be compared against predictions}

\item{metrics}{Optional list of scoring metrics to computeq}

\item{output_type_id_order}{For ordinal variables in pmf format, this is a vector of levels for pmf forecasts, in
increasing order of the levels. For all other output types, this is ignored.}
}
\value{
forecast_quantile
}
\description{
Score model output predictions against observed data
}
