% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_point_model_out.R
\name{transform_point_model_out}
\alias{transform_point_model_out}
\title{Transform mean/median model output into a forecast object}
\usage{
transform_point_model_out(model_out_tbl, target_data, output_type)
}
\arguments{
\item{model_out_tbl}{Forecast model output tibble}

\item{target_data}{Observed 'ground truth' data to be compared against forecasts}

\item{output_type}{Forecast output type: "mean" or "median"}
}
\value{
forecast_point
}
\description{
Transform mean/median model output into a forecast object
}
