% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_quantile_model_out.R
\name{transform_quantile_model_out}
\alias{transform_quantile_model_out}
\title{Transform quantile model output into a forecast object}
\usage{
transform_quantile_model_out(model_out_tbl, oracle_output)
}
\arguments{
\item{model_out_tbl}{Model output tibble with predictions}

\item{oracle_output}{Predictions that would have been generated by an oracle
model that knew the observed target data values in advance}
}
\value{
forecast_quantile
}
\description{
Transform quantile model output into a forecast object
}
